+++ varnish-sys/src/vcl/probe.rs
@@ -1,18 +1,22 @@
 use std::borrow::Cow;
-use std::ffi::c_uint;
+use std::ffi::{c_char, c_uint, CStr};
+use std::mem::size_of;
 use std::time::Duration;
 
 use serde::{Deserialize, Serialize};
 
+use crate::ffi::{vrt_backend_probe, VCL_DURATION, VCL_PROBE, VRT_BACKEND_PROBE_MAGIC};
+use crate::vcl::{IntoVCL, VclError, Workspace};
+
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub enum CowRequest<'a> {
-    URL(Cow<'a, str>),
-    Text(Cow<'a, str>),
+pub enum Request<T> {
+    URL(T),
+    Text(T),
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct CowProbe<'a> {
-    pub request: CowRequest<'a>,
+pub struct Probe<T = String> {
+    pub request: Request<T>,
     pub timeout: Duration,
     pub interval: Duration,
     pub exp_status: c_uint,
@@ -21,29 +25,14 @@ pub struct CowProbe<'a> {
     pub initial: c_uint,
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub enum Request {
-    URL(String),
-    Text(String),
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct Probe {
-    pub request: Request,
-    pub timeout: Duration,
-    pub interval: Duration,
-    pub exp_status: c_uint,
-    pub window: c_uint,
-    pub threshold: c_uint,
-    pub initial: c_uint,
-}
+pub type CowProbe<'a> = Probe<Cow<'a, str>>;
 
 impl<'a> CowProbe<'a> {
     pub fn to_owned(&self) -> Probe {
         Probe {
             request: match &self.request {
-                CowRequest::URL(cow) => Request::URL(cow.to_string()),
-                CowRequest::Text(cow) => Request::Text(cow.to_string()),
+                Request::URL(cow) => Request::URL(cow.to_string()),
+                Request::Text(cow) => Request::Text(cow.to_string()),
             },
             timeout: self.timeout,
             interval: self.interval,
@@ -52,5 +41,73 @@ impl<'a> CowProbe<'a> {
             threshold: self.threshold,
             initial: self.initial,
         }
+    }
+}
+
+/// Helper to convert a probe into a VCL object
+pub(crate) fn into_vcl_probe<T: AsRef<str>>(
+    src: Probe<T>,
+    ws: &mut Workspace,
+) -> Result<VCL_PROBE, VclError> {
+    // FIXME: this will be changed with the new Workspace API
+    #[allow(clippy::cast_ptr_alignment)]
+    let probe = unsafe {
+        ws.alloc(size_of::<vrt_backend_probe>())?
+            .as_mut_ptr()
+            .cast::<vrt_backend_probe>()
+            .as_mut()
+            .unwrap()
+    };
+
+    *probe = vrt_backend_probe {
+        magic: VRT_BACKEND_PROBE_MAGIC,
+        timeout: src.timeout.into(),
+        interval: src.interval.into(),
+        exp_status: src.exp_status,
+        window: src.window,
+        initial: src.initial,
+        ..Default::default()
+    };
+
+    match src.request {
+        Request::URL(s) => {
+            probe.url = s.as_ref().into_vcl(ws)?.0;
+        }
+        Request::Text(s) => {
+            probe.request = s.as_ref().into_vcl(ws)?.0;
+        }
+    }
+
+    Ok(VCL_PROBE(probe))
+}
+
+/// Helper to convert a VCL probe into a Rust probe wrapper
+pub(crate) fn from_vcl_probe<'a, T: From<Cow<'a, str>>>(value: VCL_PROBE) -> Option<Probe<T>> {
+    let pr = unsafe { value.0.as_ref()? };
+    assert!(
+        (pr.url.is_null() && !pr.request.is_null()) || pr.request.is_null() && !pr.url.is_null()
+    );
+    Some(Probe {
+        request: if pr.url.is_null() {
+            Request::Text(from_str(pr.request).into())
+        } else {
+            Request::URL(from_str(pr.url).into())
+        },
+        timeout: VCL_DURATION(pr.timeout).into(),
+        interval: VCL_DURATION(pr.interval).into(),
+        exp_status: pr.exp_status,
+        window: pr.window,
+        threshold: pr.threshold,
+        initial: pr.initial,
+    })
+}
+
+/// Helper function to convert a C string into a Rust string
+fn from_str<'a>(value: *const c_char) -> Cow<'a, str> {
+    if value.is_null() {
+        Cow::Borrowed("")
+    } else {
+        // FIXME: this should NOT be lossy IMO
+        unsafe { CStr::from_ptr(value).to_string_lossy() }
     }
 }